+++ CHANGELOG.rst
@@ -204,3 +204,5 @@ Changelog
 * Added option to save or discard photons arriving during the returning phase of a unidirectional scan.
 
 * Introduced ``dead_time`` parameter to a ``Movie()`` that determines the delay between volumes.
+
+* Some tests are working again.